# LF: Factorize logic to compute contract meta data