--- a/daml-lf/interpreter/src/main/scala/com/digitalasset/daml/lf/speedy/Pretty.scala
+++ b/daml-lf/interpreter/src/main/scala/com/digitalasset/daml/lf/speedy/Pretty.scala
@@ -503,12 +503,11 @@ private[lf] object Pretty {
               ) + char(
                 ']'
               )
-            case SBUCreate(ref, None) =>
-              text("$create") + char('[') + text(ref.qualifiedName.toString) + char(']')
-            case SBUCreate(ref, Some(iface)) =>
-              text("$createByInterface") + char('[') + text(ref.qualifiedName.toString) + char(
-                ','
-              ) + text(iface.qualifiedName.toString) + char(']')
+            case SBUCreate(None) =>
+              text("$create")
+            case SBUCreate(Some(iface)) =>
+              text("$createByInterface") + char(',') + text(iface.qualifiedName.toString) +
+                char(']')
             case SBUFetch(ref) =>
               text("$fetch") + char('[') + text(ref.qualifiedName.toString) + char(']')
             case SBGetTime => text("$getTime")