--- a/daml-lf/interpreter/src/main/scala/com/digitalasset/daml/lf/speedy/Profile.scala
+++ b/daml-lf/interpreter/src/main/scala/com/digitalasset/daml/lf/speedy/Profile.scala
@@ -241,8 +241,6 @@ object Profile {
       implicit val createByInterfaceDefRef: Allowed[CreateByInterfaceDefRef] = allowAll
       implicit val signatoriesDefRef: Allowed[SignatoriesDefRef] = allowAll
       implicit val observersDefRef: Allowed[ObserversDefRef] = allowAll
-      implicit val toAnyContractDefRef: Allowed[ToCachedContractDefRef] = allowAll
-      implicit val implementsDefRef: Allowed[ImplementsDefRef] = allowAll
       implicit val implementsMethodDefRef: Allowed[ImplementsMethodDefRef] = allowAll
       implicit val choiceDefRef: Allowed[ChoiceDefRef] = allowAll
       implicit val guardedChoiceDefRef: Allowed[GuardedChoiceDefRef] = allowAll
@@ -269,8 +267,6 @@ object Profile {
           case SignatoriesDefRef(tmplRef) => s"signatories @${tmplRef.qualifiedName}"
           case ObserversDefRef(tmplRef) => s"observers @${tmplRef.qualifiedName}"
           case ToCachedContractDefRef(tmplRef) => s"toAnyContract @${tmplRef.qualifiedName}"
-          case ImplementsDefRef(tmplRef, ifaceId) =>
-            s"implements @${tmplRef.qualifiedName} @${ifaceId.qualifiedName}"
           case ImplementsMethodDefRef(tmplRef, ifaceId, methodName) =>
             s"implementsMethod @${tmplRef.qualifiedName} @${ifaceId.qualifiedName} ${methodName}"
           case ChoiceDefRef(tmplRef, name) => s"exercise @${tmplRef.qualifiedName} ${name}"