between Create and Fetch.

meta data includes, signatories, observers, and contract key.

Benchmarking shows no performance impact: 

```
==> Before changes <=
Benchmark Mode  Cnt   Score   Error  Units
CollectAuthority  avgt  200  11.115 ms/op 0.025  ms/op
ReplayBenchmarkRef1  avgt   200  11.906 ms/op  0.131  ms/op
ReplayBenchmarkRef2  avgt   200   0.548 ms/op  0.004  ms/op

=> After changes <=
Benchmark Mode  Cnt   Score   Error  Units
CollectAuthority  avgt  200  11.078 ms/op  0.042 ms/op
ReplayBenchmarkRef1  avgt  200  11.976 ms/op  0.129 ms/op
ReplayBenchmarkRef2  avgt  200   0.542 ms/op  0.004 ms/op
```


CHANGELOG_BEGIN
CHANGELOG_END

### Pull Request Checklist

- [ ] Read and understand the [contribution guidelines](https://github.com/digital-asset/daml/blob/main/CONTRIBUTING.md)
- [ ] Include appropriate tests
- [ ] Set a descriptive title and thorough description
- [ ] Add a reference to the [issue this PR will solve](https://github.com/digital-asset/daml/issues), if appropriate
- [ ] Include changelog additions in one or more commit message bodies between the `CHANGELOG_BEGIN` and `CHANGELOG_END` tags
- [ ] Normal production system change, include purpose of change in description
- [ ] If you mean to change the status of a component, please make sure you keep [the Component Status page](https://github.com/digital-asset/daml/blob/main/docs/source/support/component-statuses.rst) up to date.

NOTE: CI is not automatically run on non-members pull-requests for security
reasons. The reviewer will have to comment with `/AzurePipelines run` to
trigger the build.